Move context: filter decorations into separate extension
